+++ src/cmds/keys.py
@@ -8,6 +8,35 @@ from src.util.keychain import (
     seed_from_mnemonic,
 )
 
+command_list = [
+    "generate",
+    "generate_and_print",
+    "show",
+    "add_seed",
+    "add",
+    "add_not_extended",
+    "delete",
+    "delete_all",
+]
+
+
+def help_message():
+    print("usage: chia keys command")
+    print(f"command can be any of {command_list}")
+    print("")
+    print(f"chia keys generate  (generates and adds a key to keychain)")
+    print(f"chia keys generate_and_print  (generates but does NOT add to keychain)")
+    print(f"chia keys show (displays all the keys in keychain)")
+    print(f"chia keys add_seed -m [24 words] (add a private key through the mnemonic)")
+    print(f"chia keys add -k [extended key] (add an extended private key in hex form)")
+    print(
+        f"chia keys add_not_extended -k [key] (add a not extended private key in hex form)"
+    )
+    print(
+        f"chia keys delete -f [fingerprint] (delete a key by it's pk fingerprint in hex form)"
+    )
+    print(f"chia keys delete_all (delete all private keys in keychain)")
+
 
 def make_parser(parser):
     parser.add_argument(
@@ -30,9 +59,13 @@ def make_parser(parser):
     )
 
     parser.add_argument(
-        "command", help="", type=str, nargs=1,
+        "command",
+        help=f"Command can be any one of {command_list}",
+        type=str,
+        nargs="?",
     )
     parser.set_defaults(function=handler)
+    parser.print_help = lambda self=parser: help_message()
 
 
 keychain: Keychain = Keychain()
@@ -88,8 +121,11 @@ def add_private_key(args):
     """
     Adds a private key to the keyring, without a seed (with the raw private key bytes).
     """
-
+    if args.key is None:
+        print("Please specify the key argument -k")
+        quit()
     key_hex = args.key
+    assert key_hex is not None
     extended_key = ExtendedPrivateKey.from_bytes(bytes.fromhex(key_hex))
     print(
         f"Adding private_key: {extended_key} with fingerprint {extended_key.get_public_key().get_fingerprint()}"
@@ -101,8 +137,11 @@ def add_private_key_not_extended(args):
     """
     Adds a not extended private key to the keyring. This can be used for migrating old pool sks from keys.yaml.
     """
-
+    if args.key is None:
+        print("Please specify the key argument -k")
+        quit()
     key_hex = args.key
+    assert key_hex
     private_key = PrivateKey.from_bytes(bytes.fromhex(key_hex))
     print(
         f"Adding private_key: {private_key} with fingerprint {private_key.get_public_key().get_fingerprint()}"
@@ -155,38 +194,33 @@ def delete(args):
     """
     Delete a key by it's public key fingerprint (which is an int).
     """
+    if args.fingerprint is None:
+        print("Please specify the fingerprint argument -f")
+        quit()
 
     fingerprint = args.fingerprint
+    assert fingerprint is not None
     print(f"Deleting private_key with fingerprint {fingerprint}")
     keychain.delete_key_by_fingerprint(fingerprint)
 
 
 def handler(args, parser):
+    if args.command is None or len(args.command) < 1:
+        help_message()
+        parser.exit(1)
+
     root_path: Path = args.root_path
     if not root_path.is_dir():
         raise RuntimeError(
             "Please initialize (or migrate) your config directory with chia init."
         )
 
-    command_list = [
-        "generate",
-        "generate_and_add",
-        "show",
-        "add_seed",
-        "add",
-        "add_not_extended",
-        "delete",
-        "delete_all",
-    ]
-
-    command = args.command[0]
+    command = args.command
     if command not in command_list:
-        print(f"Available commands: {command_list}")
-        return 1
+        help_message()
+        parser.exit(1)
 
     if command == "generate":
-        generate_and_print()
-    if command == "generate_and_add":
         generate_and_add()
         check_keys(root_path)
     elif command == "show":
@@ -205,3 +239,5 @@ def handler(args, parser):
         check_keys(root_path)
     elif command == "delete_all":
         keychain.delete_all_keys()
+    if command == "generate_and_print":
+        generate_and_print()